requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==3.8.0
pymongo==4.8.0
bcrypt==4.2.0
PyJWT==2.9.0
python-dotenv==1.0.1
gunicorn==23.0.0
pytest==8.3.3
